+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 - #3 added a CHANGELOG
 - #4 added a CONTRIBUTING and a SPEC
 - #5 added ability to configure based on a `.sequelizeclihelpersrc` file
+- #14: added forceConfirmation option to force user to confirm each string of each file name part